+++ test/networks/runtests.jl
@@ -1,0 +1,10 @@
+# Test internal network logic.
+
+include("./test_utils.jl")
+
+include("./iterators.jl")
+include("./cow.jl")
+include("./restrictions.jl")
+include("./topologies.jl")
+include("./nodes.jl")
+include("./edges.jl")