+++ test/user/03-components.jl
@@ -14,10 +14,13 @@ using Test
 using ..TestUser
 
 const EN = EcologicalNetworksDynamics
-import .EN: WriteError
 
 # Many small similar components tests files, although they easily diverge.
-only = [] # Only run these if specified.
+only = [
+    "./data_components/species.jl",
+    "./data_components/foodweb.jl",
+    "./data_components/body_mass.jl",
+] # Only run these if specified.
 if isempty(only)
     for subfolder in ["./data_components", "./code_components"]
         for (folder, _, files) in walkdir(joinpath(dirname(@__FILE__), subfolder))